+++ apps/backend/src/modules/activity-logs/activity-logs.ts
@@ -2,6 +2,7 @@ import { logger } from "~/libs/modules/logger/logger.js";
 import { contributorService } from "~/modules/contributors/contributors.js";
 import { gitEmailService } from "~/modules/git-emails/git-emails.js";
 import { projectApiKeyService } from "~/modules/project-api-keys/project-api-keys.js";
+import { projectService } from "~/modules/projects/projects.js";
 
 import { ActivityLogController } from "./activity-log.controller.js";
 import { ActivityLogModel } from "./activity-log.model.js";
@@ -14,6 +15,7 @@ const activityLogService = new ActivityLogService({
 	contributorService,
 	gitEmailService,
 	projectApiKeyService,
+	projectService,
 });
 const activityLogController = new ActivityLogController(
 	logger,