+++ apps/backend/src/modules/projects/projects.ts
@@ -13,4 +13,4 @@ const projectService = new ProjectService(
 );
 const projectController = new ProjectController(logger, projectService);
 
-export { projectController };
+export { projectController, projectService };