+++ apps/frontend/src/libs/constants/navigation-items.constant.ts
@@ -8,6 +8,7 @@ const SIDEBAR_ITEMS: NavigationItem[] = [
 		label: "Projects",
 		pagePermissions: [
 			PermissionKey.VIEW_ALL_PROJECTS,
+			PermissionKey.MANAGE_ALL_PROJECTS,
 			PermissionKey.VIEW_PROJECT,
 			PermissionKey.EDIT_PROJECT,
 			PermissionKey.MANAGE_PROJECT,
@@ -23,11 +24,16 @@ const SIDEBAR_ITEMS: NavigationItem[] = [
 		href: AppRoute.CONTRIBUTORS,
 		icon: "contributors",
 		label: "Contributors",
+		pagePermissions: [PermissionKey.MANAGE_ALL_PROJECTS],
 	},
 	{
 		href: AppRoute.ANALYTICS,
 		icon: "analytics",
 		label: "Analytics",
+		pagePermissions: [
+			PermissionKey.VIEW_ALL_PROJECTS,
+			PermissionKey.MANAGE_ALL_PROJECTS,
+		],
 	},
 ];