+++ apps/frontend/src/libs/helpers/helpers.ts
@@ -1,11 +1,12 @@
 export { getPermittedNavigationItems } from "./get-permitted-navigation-items/get-permitted-navigation-items.helper.js";
-export { getValidClassNames } from "./get-valid-classes.helper.js";
+export { getValidClassNames } from "./get-valid-classes/get-valid-classes.helper.js";
 export {
 	addDays,
 	checkHasPermission,
 	configureString,
 	formatDate,
 	getDifferenceInDays,
+	getRelativeDate,
 	initDebounce,
 	subtractDays,
 } from "@git-fit/shared";