+++ apps/frontend/src/pages/project-access-management/libs/components/project-group-create-form/project-group-create-form.tsx
@@ -7,36 +7,25 @@ import {
 import {
 	type ProjectGroupCreateRequestDto,
 	projectGroupCreateValidationSchema,
-	type ProjectGroupGetAllItemResponseDto,
 	actions as projectGroupsActions,
 } from "~/modules/project-groups/project-groups.js";
 
-import { filterUserProjectGroups } from "../../helpers/helpers.js";
 import { ProjectGroupForm } from "../project-group-form/project-group-form.js";
 import { DEFAULT_PROJECT_GROUP_CREATE_PAYLOAD } from "./libs/constants/constants.js";
 
 type Properties = {
 	onSubmit: (payload: ProjectGroupCreateRequestDto) => void;
-	projectGroups: ProjectGroupGetAllItemResponseDto[];
 	projectId: number;
 };
 
 const ProjectGroupCreateForm = ({
 	onSubmit,
-	projectGroups,
 	projectId,
 }: Properties): JSX.Element => {
 	const dispatch = useAppDispatch();
 
-	const { users, usersTotalCount } = useAppSelector(({ projectGroups }) => ({
-		projectGroups,
-		users: projectGroups.users,
-		usersTotalCount: projectGroups.usersTotalCount,
-	}));
-
-	const usersWithCurrentProjectGroups = filterUserProjectGroups(
-		users,
-		projectGroups,
+	const { usersTotalCount } = useAppSelector(
+		({ projectGroups }) => projectGroups,
 	);
 
 	const { page: userPage, pageSize: userPageSize } = usePagination({
@@ -61,7 +50,6 @@ const ProjectGroupCreateForm = ({
 			}}
 			onSubmit={onSubmit}
 			submitLabel="Submit"
-			users={usersWithCurrentProjectGroups}
 			validationSchema={projectGroupCreateValidationSchema}
 		/>
 	);