+++ apps/frontend/src/pages/project/project.tsx
@@ -1,6 +1,7 @@
 import {
 	Breadcrumbs,
 	Button,
+	Modal,
 	PageLayout,
 } from "~/libs/components/components.js";
 import { AppRoute, DataStatus, PermissionKey } from "~/libs/enums/enums.js";
@@ -8,14 +9,20 @@ import { checkHasPermission } from "~/libs/helpers/helpers.js";
 import {
 	useAppDispatch,
 	useAppSelector,
+	useCallback,
 	useEffect,
 	useModal,
 	useParams,
 } from "~/libs/hooks/hooks.js";
-import { actions as projectActions } from "~/modules/projects/projects.js";
+import {
+	actions as projectActions,
+	type ProjectPatchRequestDto,
+} from "~/modules/projects/projects.js";
 import { NotFound } from "~/pages/not-found/not-found.jsx";
+import { ProjectUpdateForm } from "~/pages/projects/libs/components/components.js";
 
 import {
+	ContributorsList,
 	ProjectDetailsMenu,
 	SetupAnalyticsModal,
 } from "./libs/components/components.js";
@@ -25,28 +32,84 @@ const Project = (): JSX.Element => {
 	const dispatch = useAppDispatch();
 	const { id: projectId } = useParams<{ id: string }>();
 
-	const { project, projectStatus } = useAppSelector(({ projects }) => projects);
-	const { authenticatedUser } = useAppSelector(({ auth }) => auth);
+	const {
+		project,
+		projectContributors,
+		projectContributorsStatus,
+		projectPatchStatus,
+		projectStatus,
+	} = useAppSelector(({ projects }) => projects);
 
-	const mainPermission = authenticatedUser
-		? authenticatedUser.groups.flatMap((group) => group.permissions)
-		: [];
+	const { permissionedProjectsId, projectUserPermissions, userPermissions } =
+		useAppSelector(({ auth }) => auth);
 
-	const projectPermission = authenticatedUser
-		? authenticatedUser.projectGroups.flatMap(
-				(projectGroup) => projectGroup.permissions,
-			)
-		: [];
-	const userPermissions = [...projectPermission, ...mainPermission];
+	const allPermissions = [...userPermissions, ...projectUserPermissions];
+
+	const {
+		isOpened: isSetupAnalyticsModalOpened,
+		onClose: onSetupAnalyticsModalClose,
+		onOpen: onSetupAnalyticsModalOpen,
+	} = useModal();
+
+	const {
+		isOpened: isEditModalOpen,
+		onClose: handleEditModalClose,
+		onOpen: handleEditModalOpen,
+	} = useModal();
+
+	useEffect(() => {
+		if (projectId) {
+			void dispatch(projectActions.getById({ id: projectId }));
+			void dispatch(projectActions.loadAllContributorsByProjectId(projectId));
+		}
+	}, [dispatch, projectId]);
+
+	useEffect(() => {
+		if (projectPatchStatus === DataStatus.FULFILLED) {
+			handleEditModalClose();
+
+			if (projectId) {
+				void dispatch(projectActions.getById({ id: projectId }));
+			}
+		}
+	}, [projectPatchStatus, handleEditModalClose, dispatch, projectId]);
+
+	const handleEditProject = useCallback(() => {
+		handleEditModalOpen();
+	}, [handleEditModalOpen]);
+
+	const handleProjectEditSubmit = useCallback(
+		(payload: ProjectPatchRequestDto) => {
+			if (project) {
+				void dispatch(projectActions.patch({ id: project.id, payload }));
+			}
+		},
+		[dispatch, project],
+	);
+
+	const isLoading =
+		projectStatus === DataStatus.PENDING || projectStatus === DataStatus.IDLE;
+
+	const isContributorsDataLoading =
+		projectContributorsStatus === DataStatus.PENDING ||
+		projectContributorsStatus === DataStatus.IDLE;
+
+	const isRejected = projectStatus === DataStatus.REJECTED;
+
+	const hasProject = project !== null;
+
+	const hasSetupAnalyticsPermission = checkHasPermission(
+		[PermissionKey.MANAGE_ALL_PROJECTS],
+		userPermissions,
+	);
 
 	const hasManagePermission =
 		checkHasPermission(
 			[PermissionKey.VIEW_ALL_PROJECTS, PermissionKey.MANAGE_PROJECT],
-			userPermissions,
+			allPermissions,
 		) &&
-		authenticatedUser?.projectGroups.map((project) =>
-			project.projectId.map((id) => id === project.id),
-		);
+		hasProject &&
+		permissionedProjectsId.includes(project.id);
 
 	const hasEditPermission =
 		checkHasPermission(
@@ -55,30 +118,12 @@ const Project = (): JSX.Element => {
 				PermissionKey.MANAGE_PROJECT,
 				PermissionKey.EDIT_PROJECT,
 			],
-			userPermissions,
+			allPermissions,
 		) &&
-		authenticatedUser?.projectGroups.map((project) =>
-			project.projectId.map((id) => id === project.id),
-		);
+		hasProject &&
+		permissionedProjectsId.includes(project.id);
 
-	const {
-		isOpened: isSetupAnalyticsModalOpened,
-		onClose: onSetupAnalyticsModalClose,
-		onOpen: onSetupAnalyticsModalOpen,
-	} = useModal();
-
-	useEffect(() => {
-		if (projectId) {
-			void dispatch(projectActions.getById({ id: projectId }));
-		}
-	}, [dispatch, projectId]);
-
-	const isLoading =
-		projectStatus === DataStatus.PENDING || projectStatus === DataStatus.IDLE;
-
-	const isRejected = projectStatus === DataStatus.REJECTED;
-
-	const hasProject = project !== null;
+	const hasPermission = hasManagePermission || hasEditPermission;
 
 	if (isRejected) {
 		return <NotFound />;
@@ -100,9 +145,12 @@ const Project = (): JSX.Element => {
 					<div className={styles["project-layout"]}>
 						<div className={styles["project-header"]}>
 							<h1 className={styles["title"]}>{project.name}</h1>
+
 							<ProjectDetailsMenu
-								hasPermission={hasManagePermission}
+								hasManagePermission={hasPermission}
+								onEdit={handleEditProject}
 								projectId={project.id}
+								userPermissions={userPermissions}
 							/>
 						</div>
 
@@ -114,7 +162,8 @@ const Project = (): JSX.Element => {
 								{project.description}
 							</p>
 						</div>
-						{hasEditPermission && (
+
+						{hasSetupAnalyticsPermission && (
 							<div>
 								<Button
 									label="Setup Analytics"
@@ -122,7 +171,25 @@ const Project = (): JSX.Element => {
 								/>
 							</div>
 						)}
+
+						<div className={styles["contributors-list-wrapper"]}>
+							<ContributorsList
+								contributors={projectContributors}
+								isLoading={isContributorsDataLoading}
+							/>
+						</div>
 					</div>
+
+					<Modal
+						isOpened={isEditModalOpen}
+						onClose={handleEditModalClose}
+						title="Update project"
+					>
+						<ProjectUpdateForm
+							onSubmit={handleProjectEditSubmit}
+							project={project}
+						/>
+					</Modal>
 
 					<SetupAnalyticsModal
 						isOpened={isSetupAnalyticsModalOpened}