+++ packages/shared/src/libs/enums/permission-key.enum.ts
@@ -1,5 +1,6 @@
 const PermissionKey = {
 	EDIT_PROJECT: "edit_project",
+	MANAGE_ALL_PROJECTS: "manage_all_projects",
 	MANAGE_PROJECT: "manage_project",
 	MANAGE_USER_ACCESS: "manage_user_access",
 	VIEW_ALL_PROJECTS: "view_all_projects",