+++ packages/shared/CHANGELOG.md
@@ -1,5 +1,40 @@
 # Changelog
 
+## [1.24.0](https://github.com/BinaryStudioAcademy/bsa-2024-gitfit/compare/shared-v1.23.0...shared-v1.24.0) (2024-09-18)
+
+
+### Features
+
+* add project last activity label to project list gf-274 ([#354](https://github.com/BinaryStudioAcademy/bsa-2024-gitfit/issues/354)) ([a4e7fab](https://github.com/BinaryStudioAcademy/bsa-2024-gitfit/commit/a4e7fab02b74f6d9fdbe44bd8c37b0fa22ccc5ef))
+
+## [1.23.0](https://github.com/BinaryStudioAcademy/bsa-2024-gitfit/compare/shared-v1.22.1...shared-v1.23.0) (2024-09-18)
+
+
+### Features
+
+* Add an ability to edit project group gf-342 ([#371](https://github.com/BinaryStudioAcademy/bsa-2024-gitfit/issues/371)) ([8c97374](https://github.com/BinaryStudioAcademy/bsa-2024-gitfit/commit/8c97374279cb535488bd9a197a35791e97fb7665))
+
+## [1.22.1](https://github.com/BinaryStudioAcademy/bsa-2024-gitfit/compare/shared-v1.22.0...shared-v1.22.1) (2024-09-18)
+
+
+### Bug Fixes
+
+* fix check has permission helper gf-248 ([#378](https://github.com/BinaryStudioAcademy/bsa-2024-gitfit/issues/378)) ([ffc88da](https://github.com/BinaryStudioAcademy/bsa-2024-gitfit/commit/ffc88da2ec88eb24b7e6e0dabe5f0b4ddd96a3ab))
+
+## [1.22.0](https://github.com/BinaryStudioAcademy/bsa-2024-gitfit/compare/shared-v1.21.1...shared-v1.22.0) (2024-09-17)
+
+
+### Features
+
+* add permission to manage all projects gf-248 ([#282](https://github.com/BinaryStudioAcademy/bsa-2024-gitfit/issues/282)) ([0bc48e9](https://github.com/BinaryStudioAcademy/bsa-2024-gitfit/commit/0bc48e980a5aacf3b43cab5aacf2c9583c36a3c2))
+
+## [1.21.1](https://github.com/BinaryStudioAcademy/bsa-2024-gitfit/compare/shared-v1.21.0...shared-v1.21.1) (2024-09-17)
+
+
+### Bug Fixes
+
+* improve error handling in creating activity logs gf-332 ([#365](https://github.com/BinaryStudioAcademy/bsa-2024-gitfit/issues/365)) ([9a4bb43](https://github.com/BinaryStudioAcademy/bsa-2024-gitfit/commit/9a4bb4315be768c6e5b278f546ba60736ad9b30b))
+
 ## [1.21.0](https://github.com/BinaryStudioAcademy/bsa-2024-gitfit/compare/shared-v1.20.0...shared-v1.21.0) (2024-09-17)